fix/max

+++ engine/core/Entity.hpp
@@ -2,6 +2,7 @@
 
 #include <container/Pool.hpp>
 #include <cstdint>
+#include "../Config.hpp"
 
 using namespace zet;
 
@@ -18,8 +19,8 @@ namespace mir {
 
     constexpr Id INVALID_ID{};
 #ifdef CONFIG_MAX_ENTITY
-    constexpr std::uint16_t MAX_ID = CONFIG_MAX_ENTITY;
+    constexpr std::uint32_t MAX_ID = CONFIG_MAX_ENTITY;
 #else
-    constexpr std::uint16_t MAX_ID = UINT16_MAX;
+    constexpr std::uint32_t MAX_ID = UINT32_MAX;
 #endif
 }
